% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Incorrect add numbers}
\usage{
add(...)
}
\arguments{
\item{...}{The numbers to add together.}
}
\value{
A numeric vector.
}
\description{
Incorrect add numbers
}
\examples{
add(1, 2, 3, 4)
add(c(3, 11))
}
